feat: Adding deletion of chatroom on BE& update Prisma

+++ server/src/chatrooms/chatrooms.service.ts
@@ -1,4 +1,9 @@
-import { BadRequestException, Injectable, Param } from '@nestjs/common';
+import {
+  BadRequestException,
+  Injectable,
+  Param,
+  UnauthorizedException,
+} from '@nestjs/common';
 import { Chatroom } from '@prisma/client';
 import { PrismaService } from 'src/prisma.service';
 
@@ -75,8 +80,24 @@ export class ChatroomsService {
     });
   }
 
-  delete(chatroomId: number): Promise<Chatroom> {
-    return this.prisma.chatroom.delete({ where: { id: chatroomId } });
+  async delete(chatroomId: number, userId?: number): Promise<Chatroom> {
+    try {
+      const room = await this.prisma.chatroom.findUnique({
+        where: { id: chatroomId },
+      });
+
+      if (room.userOwnerId !== userId)
+        throw new UnauthorizedException({
+          message: 'You are not the owner of this chatroom',
+          error: 'unauthorized',
+          statusCode: 401,
+        });
+
+      return this.prisma.chatroom.delete({ where: { id: chatroomId } });
+    } catch (e) {
+      console.log(e);
+      throw new BadRequestException(e);
+    }
   }
 
   async join(chatroomId: number, userId: number): Promise<Chatroom> {
